Adds support for `Enum::<T>::Variant`.
This commit changes `TypeBinding` to contain also
`prefix_type_arguments` this allows us to have both
`Enum::<T>::Variant` compiled with `prefix_type_arguments` and
`Enum::Variant::<T>` compiled with `type_arguments`. In case both are
used an error is thrown.

Errors are also thrown when `prefix_type_arguments` is used with
constants or functions. Tests included for this.

Also added error when `type_arguments` is used with constants.